auth tut #15 authorised requests

+++ client/src/components/WorkoutDetails.js
@@ -1,21 +1,29 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
+import { useAuthContext } from "../hooks/useAuthContext"
 
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
     const { dispatch } = useWorkoutsContext()
+    const { user } = useAuthContext()
 
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
         const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
 
         if (response.ok) {
-            dispatch({type: 'DELETE_WORKOUT', payload: json})
+            dispatch({ type: 'DELETE_WORKOUT', payload: json })
         }
     }
-    
+
     return (
         <div className="workout-details">
             <h4>{workout.title}</h4>
